refactor(crypto): split curve primitives
