+++ src/crypto/index.ts
@@ -1,5 +1,7 @@
 export * from './core';
-export * from './bls';
+export * from './curve_bls';
+export * from './curve_secp';
+export * from './curves';
 export * from './NUT01';
 export * from './NUT10';
 export * from './NUT11';
